backend/requirements.txt
supabase==2.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.0.0
gunicorn>=22.0.0
